bugfix: do not use hard-coded file extension .fms for formants track
In order to edit formants in the emu-webapp, the _DBconfig.json must include an
SSFFTrackDefinition named "FORMANTS". That name is hard-coded in the web app,
but the definition includes the string-typed field "fileExtension"; and that
field can be freely chosen by the user.

While fileExtension is most often set to 'fms', this need not be so. This server
will now honor the user-chosen value in that field. Up until now, it used the
hard-coded 'fms', but only for writing.

This led to the server reading from one file and writing to another file in case
something other than 'fms' was chosen by the user.